backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
audioop-lts==0.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cobble==0.1.4
cryptography==44.0.0
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.17
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
mammoth==1.9.0
markdownify==0.14.1
markitdown==0.0.1a3
numpy==2.0.2
openai==1.59.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
puremagic==1.28
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
XlsxWriter==3.2.0
youtube-transcript-api==0.6.3
